add test

+++ src/main/java/br/com/mba/inventorycontrol/repository/StoragePlaceRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StoragePlaceRepository extends JpaRepository<StoragePlace, Long> {
+public interface StoragePlaceRepository extends JpaRepository<StoragePlace, Integer> {
 
 }
